Speed up retrying of failed external link checks
We retry external link checks in a few situations (e.g., to see if the
server supports GET but not HEAD). But we were performing these retries
in a blocking manner. We now put the retries into the Typhoeus::Hydra
queue just like all the other requests, so they can be performed in
parallel. This speeds up checking 62 links on a server that does not
support HEAD from 20 seconds to 4 seconds on my machine.